+++ packages/ts-newskit/src/components/slices/shared-styles/index.ts
@@ -1,7 +1,30 @@
-import { styled, TextBlock, Block, getSizingCssFromTheme } from 'newskit';
+import {
+  styled,
+  TextBlock,
+  Block,
+  getSizingCssFromTheme,
+  getColorCssFromTheme,
+  CardLink
+} from 'newskit';
+
+export const CardHeadlineLink = styled(CardLink)<{ $color?: string }>`
+  ${({ $color }) => getColorCssFromTheme('color', $color || 'inkContrast')};
+  cursor: pointer;
+  text-decoration: none;
+
+  &&:hover,
+  &&:active {
+    text-decoration: none;
+    ${getColorCssFromTheme('color', 'interactiveLink020')};
+  }
+
+  &&:active {
+    ${getColorCssFromTheme('color', 'interactiveLink030')};
+  }
+`;
 
 export const ColouredText = styled(TextBlock)<{ $color?: string }>`
-  ${({ $color }) => $color && `color: ${$color};`};
+  ${({ $color }) => $color && getColorCssFromTheme('color', $color)};
 `;
 
 export const ContainerInline = styled(Block)`