+++ packages/ts-newskit/src/theme/times-web-light/style-presets/slices.ts
@@ -9,7 +9,7 @@ export const sliceStylePresets = {
   dashedDivider: {
     base: {
       borderStyle: 'dashed',
-      borderWidth: '{{borders.borderWidth010}}',
+      borderWidth: '{{borders.borderWidthDefault}}',
       borderColor: '{{colors.interface060}}'
     }
   }